+++ NEWS.md
@@ -2,5 +2,6 @@
 
 Still working out the basic functionality regarding the database and the app itself.
 
+* Make `cache_db_path()` smarter about the default db location via `rappdirs::user_data_dir()`.
 * Added `Dockerfile` and `docker-compose.yml` for easier deployment.
 * Added a `NEWS.md` file to track changes to the package.